Moved report converstion functions into report submodule.
