+++ components/proto/src/report/mod.rs
@@ -1,2 +1,3 @@
 pub mod messages;
 pub mod serialize;
+pub mod convert;
